Extract calendar into reusable Calendar component
